Extract runtime surface contracts
