+++ docs/DECISIONS.md
@@ -113,6 +113,7 @@
 - SSH Remote context refresh is a read-only SSH probe, not a sync layer. It reads bounded `AGENTS.md`, `.quillcode/rules.md`, `.quillcode/instructions.md`, and project `.quillcode/memories` files from the remote root into the normal project/thread context models, using hex-encoded records so paths and content with spaces survive the shell transport. Direct remote file read/write is limited to normalized project-relative paths through the SSH executor; remote apply-patch reuses the same unsafe diff-path scan and then runs `git apply --check` before `git apply` remotely. Remote hunk stage/restore reuses the git patch path-mismatch scan and `git apply --check` before applying over SSH. Remote commit trims and validates commit messages before running `git commit -m`; remote push reuses the local git-name validator for explicit refs and validates discovered current branches before pushing; remote PR creation delegates to the remote GitHub CLI after validating base/head refs; remote PR read/check/checkout/comment/review/reviewer-update/label-update/merge operations delegate to `gh pr view`, `gh pr checks`, `gh pr checkout`, `gh pr comment`, `gh pr review`, `gh pr edit`, and `gh pr merge` after validating selectors, checkout branch names, review actions, reviewer names, label names, merge methods, and body requirements; remote worktree create/remove uses the same sibling-of-project path boundary and verifies removal targets with remote `git worktree list --porcelain`. Plugin and local-environment actions still need explicit remote adapters rather than local fallbacks.
 - Execution context is shared surface metadata, not renderer guesswork. Project-bound tool cards and terminal history entries carry an optional `ExecutionContextSurface`; local contexts stay visually muted, while SSH Remote contexts render a compact chip plus a thin left rail in SwiftUI, static HTML, and the Playwright harness. This keeps old transcript scrollback clear even after the user switches projects and leaves room for future QuillCloud/relay contexts to reuse the same contract.
 - Tool-card execution-context enrichment lives in `WorkspaceExecutionContextSurfaceBuilder`. The workspace model supplies selected-project and project-list state, but the builder owns thread-project fallback, project-execution tool classification, and timeline/tool-card enrichment so future relay or remote context metadata does not spread through model mutation code.
+- Runtime issue and execution-context surface contracts live in `QuillCodeRuntimeSurface.swift`, not in the aggregate `WorkspaceSurface.swift`. `WorkspaceSurface` should keep only the composed `runtimeIssue` payload and delegate construction to runtime/execution builders, while renderers consume the shared severity, diagnostic, and execution-context records directly. Future QuillCloud relay or additional remote execution contexts should extend this contract file first instead of adding renderer-local enums.
 - Fork, compact, and cancelled-send title seeding lives in `WorkspaceThreadSeedBuilder`. The workspace model owns thread creation, selection, stores, and top-bar refresh, while the builder owns visible-message filtering, latest-turn seed selection, compact summary formatting, and first-prompt title derivation. Prompt titles split on all whitespace so cancelled whitespace-only prompts cannot create invisible chat titles.
 - SSH Remote terminal continuity uses marker metadata over the existing noninteractive shell stream. The remote wrapper captures baseline env, applies persisted terminal env deltas, runs the user command, then emits final cwd/env markers that are stripped before the entry is stored. This keeps the first implementation compatible with plain SSH while preserving `cd`, `export`, and `unset` across terminal commands.
 - Terminal session state and shell wrapping live in `WorkspaceTerminalEngine.swift`, not inline in `WorkspaceModel`. The workspace model owns async orchestration, project selection, persistence, and top-bar updates; the terminal engine owns state mutation, local/remote command wrapping, marker parsing, cwd persistence, environment deltas, and marker cleanup. This keeps terminal continuity testable as local, SSH Remote, and future relay execution contexts grow.
